refactor: 색깔 선택 뷰 패키지 정리
