+++ app/src/main/java/woowacourse/paint/presentation/ui/main/colors/ColorsAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint.presentation.ui.main
+package woowacourse.paint.presentation.ui.main.colors
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
